[Refactor] Refactoring structures
