+++ src/main/java/com/likelion/friendpass/api/matching/dto/MatchingStatusDto.java
@@ -1,5 +1,7 @@
-package com.likelion.friendpass.domain.matching;
+package com.likelion.friendpass.api.matching.dto;
 
+import com.likelion.friendpass.domain.matching.MatchingRegion;
+import com.likelion.friendpass.domain.matching.MatchingStatus;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
